better separation of stubs, remove need for symlink
