+++ stubs/phantom/action_result.py
