+++ stubs/phantom/base_connector.py
